read P01 update
read_xml function can behave depending on the Content-Type of the Header of the webpage. In case of xml content - get the xml response directly from the web page. In case of rdf Content - download the file and then read it, or just read the previously downloaded file.  Also xml_parse function is updated to read P01.